+++ src/database/index.ts
@@ -1,7 +1,7 @@
 import mongoose from "mongoose";
 
-import * as config from "../../../config.json";
-import logger from "../../handlers/logger";
+import * as config from "../../config.json";
+import logger from "../handlers/logger";
 
 export default async () => {
   await mongoose.connect(config.mongodb.url);
