+++ src/events/messageCreate/modules/counters.ts
@@ -1,4 +1,4 @@
-import counters from "../../../helpers/database/models/counterSchema";
+import counters from "../../../database/schemas/counter";
 
 import { Message } from "discord.js";
