+++ src/index.ts
@@ -1,28 +1,28 @@
 // Dependencies
 import { Client, Intents } from "discord.js"; // discord.js
 
-import database from "./helpers/database";
-import events from "./handlers/events";
-import commands from "./handlers/commands";
+import database from "./database";
 import locale from "./handlers/locale";
 import schedules from "./handlers/schedules";
 
-import config from "../config.json"; // config.json
+import events from "./handlers/events";
+import commands from "./handlers/commands";
 
-(async () => {
-  // Initialize discord.js client
-  const client = new Client({
-    intents: [
-      Intents?.FLAGS?.GUILDS,
-      Intents?.FLAGS?.GUILD_MESSAGES,
-      Intents?.FLAGS?.GUILD_MEMBERS,
-    ],
-  });
+import config from "../config.json";
 
-  await database();
-  await locale();
-  await events(client);
-  await commands(client);
-  await schedules(client);
-  await client?.login(config?.bot?.token);
-})();
+const client = new Client({
+  intents: [
+    Intents?.FLAGS?.GUILDS,
+    Intents?.FLAGS?.GUILD_MESSAGES,
+    Intents?.FLAGS?.GUILD_MEMBERS,
+  ],
+});
+
+database();
+locale();
+schedules(client);
+
+commands(client);
+events(client);
+
+client?.login(config?.bot?.token);
